* merged with trunk, rev[5139-5677]  - добален triangular в словарь вставки  - warning level = 4  - исправлена провека скролла  - подавление ворнингов  - время прогона считается более точно   * merged with dev_triangular, rev[5354-5635]    - добавление треугольного закона в рдо-процесс    - editbox'ы в диалоговом окне теперь появляются и исчезают в зависимости от выбранного закона    - изменение графического интерфеса блока create в рдо-процесс    - изменение графического интерфеса блока process в рдо-процесс    - форматирование кода    - исправлена проблема с переключением законов. Вместо ON_CBN_CLOSEUP теперь ON_CBN_SELCHANGE    - рефакторинг: переименивание функций  - перемещение
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@5684 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ rdo/app/rdo_studio/branches/dev_processgui_2/src/frame/options_ctrl.cpp
@@ -68,7 +68,7 @@ void RDOStudioFrameOptionsCtrl::setStyle( RDOStudioFrameStyle* _style )
 	lf.lfWeight    = style->theme->defaultStyle & rdoStyle::RDOStyleFont::BOLD ? FW_BOLD : FW_NORMAL;
 	lf.lfItalic    = style->theme->defaultStyle & rdoStyle::RDOStyleFont::ITALIC;
 	lf.lfUnderline = style->theme->defaultStyle & rdoStyle::RDOStyleFont::UNDERLINE;
-	lf.lfCharSet   = style->font->characterSet;
+	lf.lfCharSet   = BYTE(style->font->characterSet);
 #pragma warning(disable: 4996)
 	strcpy( lf.lfFaceName, style->font->name.c_str() );
 #pragma warning(default: 4996)
